Fix a few "assigned but unused variable" warnings
Note that "_" is the official magic name to use when you're
intentionally abandoning an assignment (typical use case, seen here:
unneeded elements being discarded in a parallel assignment) because
Ruby supresses the warning for that name.

In Ruby 2.0, you can prefix a variable name with an underscore to get
the same behavior -- without sacrificing the documentation-value of
naming things.