Fix the errant "possibly useless use of == in void context" test warnings
These are bogus -- it looks like they're triggered when the 'should =='
syntax is used before the last line of an example.

Some of these were easily fixed by just using a better matcher, e.g.
changing `a.include?(b).should == true` to `a.should.include(b)`.

In other places, though, the only option was to switch the assertion
style (i.e. `a.should == b` to `a.should.equal b`). This is kind of
gross, but it's really nice to roll with warnings enabled so we'll
notice when we accidentally add a new one in lib/.